Support transparent variant and mutually recursive types definitions
While experimenting with coqffi, it has appeared that in some cases,
being able to leverage the definition of a type introduced by a module
is desirable. For instance, consider a module to manipulate
files. This module is likely to provide a type for encoding the
various errors which can arise when interacting with a filesystem. We
could provide a set of functions to manipulate this type, but a more
idiomatic way to approach this problem is to translate the error type
into an inductive type, and to configure the extraction accordingly.