Split arguments parsing and options definition into two modules
Prior to this patch, the [Cli] module was mixing the description of
the different options which can be used to tweak coqffi output, and
the CLI argument parsing to determine which options were selected by
the user. This patch splits these two tasks. The parsing is put
directly in the file of the coqffi executable, while the list of
options are place into the coqffi library in a dedicated module.