Use coq-simple-io to handle impure primitives
This commit is based on an initial proposal by Li-yao Xia, but
generalizes the approach to avoid having to propose one option per
potential “impure backend”. Instead, we adopt a typeclass-based
approach.

We always provide an instance of the typeclass gathering primitives
interfaces for the [IO] monad of coq-simple-io.

In addition, we also introduce the -finterface, in place of -mFreeSpec
and -mITree. When -finterface is set, we generate an interface
compatible with ITree, FreeSpec and potentially other frameworks based
on a similar free monad. Besides, we generate helpers for said
frameworks to use this interface easily.  It is up to FreeSpec and
ITree to provide the expected instance on [Inject] for them to remain
or become compatible.