Support older OCaml versions up to 4.08
In order to support various version of the OCaml, we propose the
following approach: the main branch of coqffi is devoted to the latest
OCaml version, and in the `patches` directory, we provide the
necessary changes to apply to make it build with another version of an
older version.

This may seem cumbersome to use compare to a branch-based approach,
but we anticipate it will greatly ease the cost to backport new
features for older OCaml compiler.