Add the [IntFacts] module to reason with [i63] terms
This patch introduces a new module with a collection of axioms and
lemmas about the [i63] type. It is not complete, and several axioms
can probably be turned into proper lemmas at some point, but it has
the benefit to provide useful tools to reason about integer
arithmetic. In particular, it has been battle-tested to reason about
absence of integers overflows.