Introduce the `freespec' feature to generate FreeSpec semantics
The starting point of coqffi was to be able to easily generate
FreeSpec’s semantics. When we decided to expand its purpose to be a
more general-purpose tool that the Coq community could use, we also
decided to remove the code too specific to FreeSpec.

Now that we have a convenient enough way to deal with features (that
can be enable or disable at will), we think it is a good moment to
reintroduce this lost code.