Refactor how the Coq module is generated
We introduce the [Vernac] module as an additional level of
abstraction.  Now, coqffi attempts to construct a list of each
vernacular commands it has to generate prior to actually generate the
module.  There are two advantages to proceeds this way.  First, we
centralize the pretty-printing code related to vernacular commands
(whereas before, several functions were generated the same kind of
commands).  Second, if we encounter an error during the computation of
the commands to generate, we can exit prior to generating something.

Whether or not this refactoring makes the code easier to read remains
to be seen.