Support the [@@may_raise] annotation
If a pure function or an OCaml value is susceptible to throw an
exception, then coqffi users can marked them using the [@@may_raise]
annotation. In such a case, axiomatic definitions generated by coqffi
will have their codomain lifted in the [sum] type of Coq, which is
analoguous to [result] in OCaml, and the extraction mechanism of Coq
is configured accordingly.