Support aliases for Coq reserved keywords and OCaml operators
Operators (i.e., entries of the form [val (+) : t -> t -> t]) are now
supported correctly by coqffi. As of now, arithmetic operators are
supported. Similarly, the introduced alias mechanism allows for
dealing with reserved keywords in Gallian, like [return].

The list provided in the [Alias] module is minimal, and should be
extended prior to the 1.0.0 release.