Support polymorphic constructors
Prior to this patch, types such as

    type box = Box : 'a -> box

were not supported, and could even be used to produce ill-formed Coq
modules. For instance, consider the following snippet

    type t
    type u = U : 't -> u

In presence of such pattern, coqffi was considering that the ['t]
polymorphic type argument of [U] was actually the type [t] previously
introduced.

This patch fixes this issue, and we take the opportunity to refactor
the [variant_entry] and [exception_entry] types to make them use
[prototype_repr]. This simplifies the code in the [Vernac] module, as
the latter was actually constructing values of type [prototype_repr]
itself.