Fix the order of type declaration
Prior to this patch, it could happen that coqffi would produce a Coq
module where it would have turn opaque a transparent types it knows
how to translate. This was due to the search for mutually recursive
definitions, which mingled the input. Note that, when the translation
table was computed ahead of time, the generated Coq module was
definitely ill-formed.

We fix this issue by modifying the DFS algorithm to remember the
position of each node within the input list. This allows us to sort
the result of the DFS once it is terminated.