Add support for Lwt asynchronous function
With this patch, we add a new feature to coqffi: lwt.  When this
feature is enabled, coqffi starts treating functions computing
promises differently. First, it gathers them into a dedicated
typeclass, with an instance for the [Lwt.t] type (provided by the
user). Also, it generates an instance for the “synchronous” typeclass,
by lifting synchronous functions using [Lwt.return]. This means the
implementation of the synchronous typeclass remains blocking.
Finally, an interface is also generated for the asynchronous
primitives. A semantics could be generated by leveraging
[Lwt_main.run], but it is out of the scope of this patch.