Add a feature to generate “witness files”
A witness file in this context is an auxiliary file which contains a
s-expression whose purpose is to summarize the types introduced by the
generated Coq module. Its purpose is to be used in a next call of
`coqffi' to extend its default translation tables. This feature will
be implemented in a later patch.

An example of a generated witness file is

    (translations
      (Examples.Polytype.foo . Test.foo)
      (Examples.Polytype.llist . Test.llist)
      (Examples.Polytype.moo . Test.moo)
      (Examples.Polytype.gadt . Test.gadt)
      (Examples.Polytype.m . Test.m)
      (Examples.Polytype.box . Test.box))