Consume witness files to populate translation tables
This allows to deal with dependencies between MLI files, albeit in a
very cumbersome way. The key idea is the following: if a module M2
depends on a module M1 (meaning M2 uses types introduced by M1), then
when generating a FFI module for M1 we use the `-fwitness' feature to
generate `M1.coqffi' in addition to `M1.v'. Then, when generating a
FFI module for M2, we use the `-I' option with a path to `M1.coqffi',
and `-r' with the namespace of the module.

You can have a look at the dune rules for `M1.v' and `M2.v' in
`examples/src/dune'.