Improve support for shadowing types of the same module
Prior to this patch, the following snippet would confused coqffi

    type t
    module T : sig
        type key = t
        type 'a t = 'a option
    end

This is no longer the case with this patch. Prior to doing a DFS on
type definition to find cycles, we pre-segment the list of type entry
using the [rec_status] field.

By doing this, we allow ourselves to treat types such as

    type t = 'a u
    and 'a u = 'a option

as not mutually recursive.