Fix the generation of Coq record
The previous patch was not perfect, far from it even. This patch fixes
the main issues, and ensures everything works by using records in one
of the binary examples.

This is another proof, if needed, that the current way we test coqffi
is not enough (and that compiling a Coq module using Extraction
commands is not remotely a guarantee that we are doing something
well).