Improve coqffi resilience against unfortunate shadowing in FFI
OCaml is more permissive than Coq when it comes to naming things. For
instance, in OCaml you can have a type [t] and a value [t] in the same
module, sometimes with the following pattern

    val t : t

Another example is that you can have several records with the same
field name.  And the list goes on.

In Coq, introducing something of the form

    Axiom t : t.

will work, but with the unfortunate benefit of shadowing the type [t].
This patch introduces a significant refactoring to deal with these
refactoring, including with constructors and fields.

Unfortunately, contrary to the author initial ambition, this patch
does *not* resolve the question of shadowing once and for all. It is
possible that a helper introduced by coqffi can shadow another helper.

For instance, consider the following mli

    val fooBar : 'foo -> 'bar Lwt.t
    exception FooBar

if the `lwt' and `interface' features are enabled, then coqffi will
generate an interface type with a constructor [FooBar] *and* an
inductive type [FooBar] for the exception.

These tricky corner cases are left as an exercise for the reader.