Fix support for labelled and optional arguments
The previous patch introduces a number of issues that we needed be
fixed.  This is achieved with the present patch.  One issue remaining
is the fact that there might be a collusion between named and
positioned arguments, more precisely that if a named argument is of
the form `x[0-9]+`, it is likely that an issue will arise.